Update tool bar visuals and add them to all necessary places
