+++ .github/CREDITS.md
@@ -1,2 +1,4 @@
 https://sketchfab.com/3d-models/inside-galaxy-skybox-hdri-360-panorama-96183eed1ae2425baddb75ca92103f87
 https://medium.com/@ellisande_poet/subscriptions-with-react-hooks-d329d473efc8
+https://github.com/pickles976/GalaxyThreeJS
+https://icons.getbootstrap.com/#icons
