--- src/features/procedural-galaxy-creator/components/creator/index.ts
@@ -1,4 +1,0 @@
-export * from "./base-factory";
-export * from "./base-pipeline";
-export * from "./base-step";
-export * from "./inputs";
