+++ src/features/procedural-galaxy-creator/components/step-config-tables.ts
